Put tutorial steps in own file - #1033
